+++ Components/us_measurement.h
@@ -5,11 +5,11 @@ class US_Measurement {
 
     public:
     int servo_angle;
-    float distance_measured;
+    unsigned long distance_measured;
     US_Measurement();
-    US_Measurement(int servo_angle, float distance_measured);
+    US_Measurement(int servo_angle, unsigned long distance_measured);
     void set_servo_angle(int servo_angle);
-    void set_distance_measured(float distance_measured);
+    void set_distance_measured(unsigned long distance_measured);
 
     ~US_Measurement();
 };
